+++ src/context/kemaisession.h
@@ -4,8 +4,10 @@
 
 #include <QObject>
 
+#include "client/kimaiCache.h"
 #include "client/kimaiapi.h"
 #include "client/kimaiclient.h"
+#include "monitor/kimaiEventsMonitor.h"
 
 namespace kemai {
 
@@ -14,16 +16,18 @@ class KemaiSession : public QObject
     Q_OBJECT
 
 public:
-    KemaiSession(std::shared_ptr<KimaiClient> kimaiClient);
+    KemaiSession(const std::shared_ptr<KimaiClient>& kimaiClient);
     ~KemaiSession() override;
 
     std::shared_ptr<KimaiClient> client() const;
+    const KimaiCache& cache() const;
 
     /*
      * send some request to identify instance
      */
     void refreshSessionInfos();
     void refreshCurrentTimeSheet();
+    void refreshCache(const std::set<KimaiCache::Category>& categories = {});
 
     bool hasPlugin(ApiPlugin apiPlugin) const;
     User me() const;
@@ -45,15 +49,13 @@ private:
     void requestTimeSheetConfig();
     void requestPlugins();
 
-    void setCurrentTimeSheet(const TimeSheet& timeSheet);
-    void clearCurrentTimeSheet();
-
     std::shared_ptr<KimaiClient> mKimaiClient;
+    KimaiCache mKimaiCache;
+    KimaiEventsMonitor mKimaiMonitor;
     QVersionNumber mKimaiVersion;
     Plugins mPlugins;
     User mMe;
     TimeSheetConfig mTimeSheetConfig;
-    std::optional<TimeSheet> mCurrentTimeSheet;
 };
 
 } // namespace kemai