+++ CHANGELOG.md
@@ -6,6 +6,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+## [0.9.0] - 2023-03-27
+
+Special thanks to @shrippen for its support through [sponsorship](https://github.com/sponsors/AlexandrePTJ).
+
+### Added
+- Periodic check of current timesheet [#80](https://github.com/AlexandrePTJ/kemai/issues/80).
+
+### Fixed
+- Regression resetting dropdown fields when stopping timesheet [#77](https://github.com/AlexandrePTJ/kemai/issues/77).
+
+### Changed
+- Allow selection of project without first selecting customer [#74](https://github.com/AlexandrePTJ/kemai/issues/74).
+- Improve Loading of Customer / Projects / Activities [#44](https://github.com/AlexandrePTJ/kemai/issues/44).
+- Improve filter on Customer / Projects / Activities [#75](https://github.com/AlexandrePTJ/kemai/issues/75).
+
+
+## [0.8.0] - 2023-03-14
+
 ### Added
 - TimeSheet description and tags are saved on stop [#43](https://github.com/AlexandrePTJ/kemai/issues/43).
 - Option to enable/disable Kemai's update check [#68](https://github.com/AlexandrePTJ/kemai/issues/68).
@@ -115,7 +134,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial version.
 
 
-[Unreleased]: https://github.com/AlexandrePTJ/kemai/compare/0.7.1...HEAD
+[Unreleased]: https://github.com/AlexandrePTJ/kemai/compare/0.9.0...HEAD
+[0.9.0]: https://github.com/AlexandrePTJ/kemai/compare/0.8.0...0.9.0
+[0.8.0]: https://github.com/AlexandrePTJ/kemai/compare/0.7.1...0.8.0
 [0.7.1]: https://github.com/AlexandrePTJ/kemai/compare/0.7.0...0.7.1
 [0.7.0]: https://github.com/AlexandrePTJ/kemai/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/AlexandrePTJ/kemai/compare/0.5.0...0.6.0